+++ packages/svelte-vscode/src/utils.ts
@@ -7,3 +7,15 @@ export function btoa(decoded: string) {
     const buffer = Buffer.from(decoded, 'utf8');
     return buffer.toString('base64');
 }
+
+export function debounce<T>(
+    fn: (...args: T[]) => void,
+    miliseconds: number
+): (...args: T[]) => void {
+    let timeout: ReturnType<typeof setTimeout>;
+
+    return (...args: T[]) => {
+        clearTimeout(timeout);
+        timeout = setTimeout(() => fn(...args), miliseconds);
+    };
+}